refactor: moved ws server to separate file
